answering Philippe's comments
