+++ scripts/scil_get_subset_streamlines.py
@@ -3,13 +3,13 @@
 import argparse
 
 from dipy.io.streamline import save_tractogram
-from scilpy.io.streamlines import load_tractogram_with_reference
 
+from scilpy.io.streamlines import load_tractogram_with_reference
+from scilpy.io.utils import (add_overwrite_arg,
+                             add_reference_arg,
+                             assert_inputs_exist,
+                             assert_outputs_exist)
 from scilpy.tracking.tools import get_subset_streamlines
-from scilpy.io.utils import (assert_inputs_exist,
-                             assert_outputs_exist,
-                             add_overwrite_arg,
-                             add_reference_arg)
 
 
 def _build_args_parser():
